Final File Cleanup
Files cleaned. Only the Home page and Settings page are functional.
