+++ app/lib/screens/favorite/favorite_screen.dart
@@ -10,7 +10,7 @@ import 'package:flutter/material.dart';
 import 'package:data_table_2/data_table_2.dart';
 import 'package:intl/intl.dart';
 import 'package:profit_taker_analyzer/main.dart';
-import 'package:profit_taker_analyzer/services/last_runs.dart';
+// import 'package:profit_taker_analyzer/services/last_runs.dart';
 import 'package:profit_taker_analyzer/utils/utils.dart';
 import 'package:profit_taker_analyzer/widgets/dialogs.dart';
 import 'package:profit_taker_analyzer/widgets/text_widgets.dart';
@@ -191,7 +191,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
     runDataList.clear(); // Clear the existing data in runDataList
 
     // Fetch new data
-    allRuns = getFavoritedRuns();
+    // allRuns = getFavoritedRuns();
 
     // Call getRunDetails with the fresh list of RunData objects
     getRunDetails(allRuns, allRuns.length, runDataList);
