+++ app/lib/screens/home/ui/error_view.dart
