+++ app/lib/screens/home/widgets/run_analysis/compact_view/compact_phase_card.dart
